Authentication and Authorization

+++ app/Http/Controllers/ArtistController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Artist;
 use App\Models\Song;
+use Illuminate\Support\Facades\Gate;
 
 class ArtistController extends Controller
 {
@@ -13,10 +14,13 @@ class ArtistController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Artist::class);
+
         $artists = Artist::get();
+
         return view('artists.index', [
             'artists' => $artists
-        ],);
+        ], );
     }
 
     /**
@@ -24,6 +28,8 @@ class ArtistController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Artist::class);
+
         return view('artists.create');
     }
 
@@ -32,14 +38,18 @@ class ArtistController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Artist::class);
+
+        $request->validate([
+            'name' => ['required', 'string', 'min:3', 'max:255']
+        ]);
+
         $artist_name = $request->get('name');
-        if ($artist_name == null) {
-            return redirect()->back()->with('error', 'Artist name is required');
-        }
-        
+
         $artist = new Artist();
         $artist->name = $artist_name;
         $artist->save();
+
         return redirect()->route('artists.index');
     }
 
@@ -56,9 +66,11 @@ class ArtistController extends Controller
 
     public function show(Artist $artist)
     {
+        Gate::authorize('view', $artist);
+
         return view('artists.show', [
             'artist' => $artist,
-        ],);
+        ], );
     }
 
     /**
@@ -66,9 +78,11 @@ class ArtistController extends Controller
      */
     public function edit(Artist $artist)
     {
+        Gate::authorize('update', $artist);
+
         return view('artists.edit', [
             'artist' => $artist,
-        ],);
+        ], );
     }
 
     /**
@@ -76,6 +90,12 @@ class ArtistController extends Controller
      */
     public function update(Request $request, Artist $artist)
     {
+        Gate::authorize('update', $artist);
+
+        $request->validate([
+            'name' => ['required', 'string', 'min:3', 'max:255'],
+        ],);
+
         $artist->name = $request->get('name');
         $artist->save();
         return redirect()->route('artists.show', ['artist' => $artist]);
@@ -86,6 +106,8 @@ class ArtistController extends Controller
      */
     public function destroy(Artist $artist)
     {
+        Gate::authorize('delete', $artist);
+
         if ($artist->songs->isEmpty()) {
             $artist->delete();
             return redirect()->route('artists.index');
@@ -93,13 +115,24 @@ class ArtistController extends Controller
         return redirect()->back()->with('error', 'Artist cannot be deleted because it has songs');
     }
 
-    public function createSong(Artist $artist) {
+    public function createSong(Artist $artist)
+    {
+        Gate::authorize('update', $artist);
+
         return view('artists.create-song', [
             'artist' => $artist,
-        ],);
+        ], );
     }
-    
-    public function storeSong(Request $request, Artist $artist) {
+
+    public function storeSong(Request $request, Artist $artist)
+    {
+        Gate::authorize('update', $artist);
+
+        $request->validate([
+            'title' => ['required', 'string', 'min:4', 'max:255'],
+            'duration' => ['required', 'integer', 'min:10'],
+        ],);
+
         $song = new Song(); // use App\Models\Song;
         $song->title = $request->get('title');
         $song->duration = $request->get('duration');
